OpenGL/misc/requirements.txt
PyOpenGL>=3.1.6
glfw>=2.5.5
PyGLM>=2.7.0
pillow>=9.4.0